pyproject.toml
[tool.poetry]
name = "backend-finmanagment"
version = "0.1.0"
description = "Backend of the financial managment api"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.115.12"}
python-dotenv = "^1.1.0"
pytest = "^8.3.5"
ruff = "^0.11.11"
sqlalchemy = "^2.0.41"
alembic = "^1.16.1"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = "^1.7.4"
pymysql = "^1.1.1"
aiomysql = "^0.2.0"
aiosqlite = "^0.21.0"
bcrypt = "^4.3.0"
fastapi-utils = "^0.8.0"
fastapi-mail = "^1.5.0"
typing-inspect = "^0.9.0"
redis = "^4.6.0"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^1.0.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
pythonpath = ["src"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"